fix: harden Codex reset credit display
