+++ Sources/CodexBar/MenuCardView.swift
@@ -119,8 +119,8 @@ struct UsageMenuCardView: View {
         let creditsRemaining: Double?
         let creditsHintText: String?
         let creditsHintCopyText: String?
-        var codexResetCreditsText: String? = nil
-        var codexResetCreditsDetailText: String? = nil
+        var codexResetCreditsText: String?
+        var codexResetCreditsDetailText: String?
         let providerCost: ProviderCostSection?
         let tokenUsage: TokenUsageSection?
         let placeholder: String?
@@ -578,7 +578,9 @@ struct UsageMenuCardUsageSectionView: View {
                 InlineUsageDashboardContent(model: dashboard)
             } else if !liveModel.usageNotes.isEmpty {
                 UsageNotesContent(notes: liveModel.usageNotes)
-            } else if let placeholder = liveModel.placeholder, liveModel.metrics.isEmpty, liveModel.codexResetCreditsText == nil {
+            } else if let placeholder = liveModel.placeholder, liveModel.metrics.isEmpty,
+                      liveModel.codexResetCreditsText == nil
+            {
                 Text(placeholder)
                     .foregroundStyle(MenuHighlightStyle.secondary(self.isHighlighted))
                     .font(.subheadline)
@@ -631,41 +633,6 @@ struct UsageMenuCardCreditsSectionView: View {
     private var liveModel: UsageMenuCardView.Model {
         guard self.model.usesLiveSubtitle else { return self.model }
         return self.refreshMonitor?.model(for: self.model.provider, fallback: self.model) ?? self.model
-    }
-}
-
-private struct CodexResetCreditsContent: View {
-    let text: String
-    let detailText: String?
-    @Environment(\.menuItemHighlighted) private var isHighlighted
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            HStack(alignment: .firstTextBaseline) {
-                Text(L("Limit Reset Credits"))
-                    .font(.body)
-                    .fontWeight(.medium)
-                    .lineLimit(1)
-                Spacer()
-                Text(self.text)
-                    .font(.footnote)
-                    .foregroundStyle(MenuHighlightStyle.secondary(self.isHighlighted))
-                    .lineLimit(1)
-            }
-            if let detailText, !detailText.isEmpty {
-                Text(detailText)
-                    .font(.footnote)
-                    .foregroundStyle(MenuHighlightStyle.secondary(self.isHighlighted))
-                    .lineLimit(1)
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel([
-            L("Limit Reset Credits"),
-            self.text,
-            self.detailText,
-        ].compactMap(\.self).joined(separator: ", "))
     }
 }
 
@@ -1024,40 +991,6 @@ extension UsageMenuCardView.Model {
             notes.append(L("API key limit unavailable right now"))
         }
         return notes + subscriptionNotes
-    }
-
-    private static func codexResetCreditsText(input: Input) -> String? {
-        guard input.provider == .codex,
-              input.showOptionalCreditsAndExtraUsage,
-              let resetCredits = input.snapshot?.codexResetCredits,
-              resetCredits.availableCount > 0
-        else {
-            return nil
-        }
-        let count = resetCredits.availableCount
-        if count == 1 {
-            return L("1 manual reset available")
-        }
-        return String(format: L("%d manual resets available"), count)
-    }
-
-    private static func codexResetCreditsDetailText(input: Input) -> String? {
-        guard input.provider == .codex,
-              input.showOptionalCreditsAndExtraUsage,
-              let resetCredits = input.snapshot?.codexResetCredits,
-              let expiresAt = resetCredits.nextExpiringAvailableCredit?.expiresAt
-        else {
-            return nil
-        }
-        let timeText: String
-        switch input.resetTimeDisplayStyle {
-        case .absolute:
-            timeText = UsageFormatter.resetDescription(from: expiresAt, now: input.now)
-        case .countdown:
-            let countdown = UsageFormatter.resetCountdownDescription(from: expiresAt, now: input.now)
-            timeText = countdown == "now" ? L("now") : countdown
-        }
-        return String(format: L("Next expires %@"), timeText)
     }
 
     private static func openRouterSpendNotes(_ usage: OpenRouterUsageSnapshot) -> [String] {
