+++ Sources/CodexBar/Resources/ar.lproj/Localizable.strings
@@ -1087,3 +1087,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "أرصدة إعادة تعيين الحد";
+"1 manual reset available" = "تتوفر إعادة تعيين يدوية واحدة";
+"%d manual resets available" = "تتوفر %d عمليات إعادة تعيين يدوية";
+"Next expires %@" = "تنتهي صلاحية التالية %@";
