+++ Sources/CodexBar/Resources/ca.lproj/Localizable.strings
@@ -942,3 +942,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "Crèdits de restabliment del límit";
+"1 manual reset available" = "Hi ha 1 restabliment manual disponible";
+"%d manual resets available" = "Hi ha %d restabliments manuals disponibles";
+"Next expires %@" = "El següent caduca %@";
