+++ Sources/CodexBar/Resources/es.lproj/Localizable.strings
@@ -942,3 +942,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "Créditos para restablecer límites";
+"1 manual reset available" = "1 restablecimiento manual disponible";
+"%d manual resets available" = "%d restablecimientos manuales disponibles";
+"Next expires %@" = "El siguiente caduca %@";
