+++ Sources/CodexBar/Resources/id.lproj/Localizable.strings
@@ -1086,3 +1086,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "Kredit pengaturan ulang batas";
+"1 manual reset available" = "1 pengaturan ulang manual tersedia";
+"%d manual resets available" = "%d pengaturan ulang manual tersedia";
+"Next expires %@" = "Berikutnya kedaluwarsa %@";
