+++ Sources/CodexBar/Resources/ko.lproj/Localizable.strings
@@ -1053,3 +1053,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "한도 재설정 크레딧";
+"1 manual reset available" = "수동 재설정 1회 사용 가능";
+"%d manual resets available" = "수동 재설정 %d회 사용 가능";
+"Next expires %@" = "다음 만료: %@";
