+++ Sources/CodexBar/Resources/pl.lproj/Localizable.strings
@@ -1086,3 +1086,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "Kredyty resetowania limitu";
+"1 manual reset available" = "Dostępny 1 ręczny reset";
+"%d manual resets available" = "Dostępne ręczne resety: %d";
+"Next expires %@" = "Następny wygasa %@";
