+++ Sources/CodexBar/Resources/sv.lproj/Localizable.strings
@@ -1081,3 +1081,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "Krediter för gränsåterställning";
+"1 manual reset available" = "1 manuell återställning tillgänglig";
+"%d manual resets available" = "%d manuella återställningar tillgängliga";
+"Next expires %@" = "Nästa upphör %@";
