+++ Sources/CodexBar/Resources/th.lproj/Localizable.strings
@@ -1087,3 +1087,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "เครดิตรีเซ็ตขีดจำกัด";
+"1 manual reset available" = "มีการรีเซ็ตด้วยตนเอง 1 ครั้ง";
+"%d manual resets available" = "มีการรีเซ็ตด้วยตนเอง %d ครั้ง";
+"Next expires %@" = "รายการถัดไปหมดอายุ %@";
