+++ Sources/CodexBar/Resources/tr.lproj/Localizable.strings
@@ -1084,3 +1084,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "Limit Sıfırlama Kredileri";
+"1 manual reset available" = "1 manuel sıfırlama kullanılabilir";
+"%d manual resets available" = "%d manuel sıfırlama kullanılabilir";
+"Next expires %@" = "Sonraki sona erme %@";
