+++ Sources/CodexBar/Resources/vi.lproj/Localizable.strings
@@ -1083,3 +1083,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "Lượt đặt lại giới hạn";
+"1 manual reset available" = "Có 1 lần đặt lại thủ công";
+"%d manual resets available" = "Có %d lần đặt lại thủ công";
+"Next expires %@" = "Lượt tiếp theo hết hạn %@";
