+++ Sources/CodexBar/Resources/zh-Hant.lproj/Localizable.strings
@@ -955,3 +955,7 @@
 "language_arabic" = "العربية";
 "language_persian" = "فارسی";
 "language_thai" = "ไทย";
+"Limit Reset Credits" = "限額重設額度";
+"1 manual reset available" = "可使用 1 次手動重設";
+"%d manual resets available" = "可使用 %d 次手動重設";
+"Next expires %@" = "下一個將於 %@ 到期";
