+++ Sources/CodexBarCore/Providers/Codex/CodexProviderDescriptor.swift
@@ -252,13 +252,12 @@ struct CodexOAuthFetchStrategy: ProviderFetchStrategy {
                 sourceLabel: "oauth")
         }
 
-        guard let credits else {
+        guard credits != nil || (resetCredits?.availableCount ?? 0) > 0 else {
             throw UsageError.noRateLimitsFound
         }
 
-        // Credits can still be useful when the OAuth API omits or partially
-        // fails to decode rate-limit windows. Returning the partial OAuth result
-        // prevents auto mode from escalating a usable response into CLI fallback.
+        // Credit balances and manual resets remain useful when OAuth omits
+        // rate-limit windows. Keep the partial result instead of discarding it.
         return CodexOAuthFetchStrategy().makeResult(
             usage: UsageSnapshot(
                 primary: nil,
@@ -284,11 +283,13 @@ extension CodexOAuthFetchStrategy {
     static func _mapResultForTesting(
         _ data: Data,
         credentials: CodexOAuthCredentials,
+        resetCredits: CodexRateLimitResetCreditsSnapshot? = nil,
         sourceMode: ProviderSourceMode = .oauth) throws -> ProviderFetchResult
     {
         let usageResponse = try JSONDecoder().decode(CodexUsageResponse.self, from: data)
         return try Self.makeResult(
             usageResponse: usageResponse,
+            resetCredits: resetCredits,
             credentials: credentials,
             updatedAt: Date(),
             sourceMode: sourceMode)
